+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- Added patch of type `UpdateEntity` to update multiple fields (without arguments) [#6](https://github.com/dividab/gql-cache-patch/issues/6).
+
 ## [v0.12.0] - 2019-04-01
 
 ### Added